+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## v0.1
 
-### v0.1.1 *2/1/25 4pm
+### v0.1.2 *2/1/25 1 hr*
+
+- set up player info and gardenState object
+- movement events on WASD and arrow keys, default speed 5
+- update loop syncs all garden information across clients
+
+### v0.1.1 *2/1/25 45 mins*
 
 - glitch init from super auto lich remix
 - added readme and TODO section